requirements.txt
requests==2.33.1
pytest==9.0.3
pytest-xdist==3.8.0
allure-pytest==2.16.0
python-dotenv==1.2.2
loguru==0.7.3